+++ aws-sdk-core/lib/aws-sdk-core/machinelearning.rb
@@ -1,0 +1,5 @@
+Aws.add_service(:MachineLearning, {
+  api: "#{Aws::API_DIR}/machinelearning/2014-12-12/api-2.json",
+  docs: "#{Aws::API_DIR}/machinelearning/2014-12-12/docs-2.json",
+  paginators: "#{Aws::API_DIR}/machinelearning/2014-12-12/paginators-1.json",
+})